[TEC-3339] Fix namespace of Api module and env-variables to FlowcommerceSpree
Fix broken Spree::Variant.flow_import_item method. It is now saving all the imported data - the common part being differentiated from the local part, belonging to current experience.
