+++ app/models/spree/gateway/spree_flow_gateway.rb
@@ -99,7 +99,7 @@ module Spree
         # tokenize with Flow
         # rescue Io::Flow::V0::HttpClient::ServerError
         card_form = ::Io::Flow::V0::Models::CardForm.new(data)
-        result    = FlowCommerce.instance.cards.post(::Flow::ORGANIZATION, card_form)
+        result    = FlowCommerce.instance.cards.post(::FlowcommerceSpree::ORGANIZATION, card_form)
 
         @credit_card.update_column :gateway_customer_profile_id, result.token
       end
