+++ app/views/spree/admin/shared/_order_summary_flow.html.erb
@@ -6,7 +6,7 @@
 
 <% if @order.flow_data['order'] %>
   <p>Experience: <b><%= @order.flow_data.dig('order', 'experience', 'key') %></b></p>
-  <a target="_fc_<%= @order.id %>" href="https://console.flow.io/<%= Flow::ORGANIZATION %>/orders/<%= @order.number %>">Open in Flow console</a>
+  <a target="_fc_<%= @order.id %>" href="https://console.flow.io/<%= FlowcommerceSpree::ORGANIZATION %>/orders/<%= @order.number %>">Open in Flow console</a>
 <% else %>
   <p>Order is not localized with Flow.</p>
 <% end %>
