+++ lib/flow/order.rb
@@ -159,7 +159,7 @@ class Flow::Order
     flow_number = @order.flow_number
 
     opts = {}
-    opts[:organization] = Flow::ORGANIZATION
+    opts[:organization] = FlowcommerceSpree::ORGANIZATION
     opts[:experience]   = @experience.key
     opts[:expand]       = 'experience'
 
@@ -204,18 +204,18 @@ class Flow::Order
     @use_get = false unless @order.flow_data['order']
 
     if @use_get
-      @response ||= Flow.api :get, '/:organization/orders/%s' % @body[:number], expand: 'experience'
+      @response ||= FlowcommerceSpree::Api.run :get, '/:organization/orders/%s' % @body[:number], expand: 'experience'
     else
       # replace when fixed integer error
       # @body[:items].map! { |item| ::Io::Flow::V0::Models::LineItemForm.new(item) }
       # opts[:experience] = @experience.key
       # order_put_form = ::Io::Flow::V0::Models::OrderPutForm.new(@body)
-      # r FlowCommerce.instance.orders.put_by_number(Flow::ORGANIZATION, @order.flow_number, order_put_form, opts)
+      # r FlowCommerce.instance.orders.put_by_number(FlowcommerceSpree::ORGANIZATION, @order.flow_number, order_put_form, opts)
 
       # cache last order/put for debug purposes
       FlowSettings.set 'flow-order-put-body-%s' % @body[:number], @body.to_json
 
-      @response = Flow.api :put, '/:organization/orders/%s' % @body[:number], opts, @body
+      @response = FlowcommerceSpree::Api.run :put, '/:organization/orders/%s' % @body[:number], opts, @body
     end
   end
