+++ lib/flowcommerce_spree/engine.rb
@@ -1,4 +1,10 @@
 module FlowcommerceSpree
+  ORGANIZATION = ENV.fetch('FLOW_ORGANIZATION', 'flow.io')
+  BASE_COUNTRY = ENV.fetch('FLOW_BASE_COUNTRY', 'USA')
+  API_KEY = ENV.fetch('FLOW_API_KEY', 'test_key')
+
+  ENV['FLOW_TOKEN'] = API_KEY
+
   class Engine < ::Rails::Engine
     require 'spree/core'
     isolate_namespace FlowcommerceSpree
